Don't strip pathnames matching /blib/ in the PathTools Makefile.PL
PathTools was brute force removing any @INC entries that matched /blib/
from @INC - ie the substring blib, not even just directories named blib.
This prevents blead from building on any path that happens to contain the
substring.

The code in question seems to have been in the Makefile.PL since 2004, and is
not present in any other distribution on CPAN, so the most likely explanation
is that it's debugging code left in by accident.

Hence remove it.