Smarter handling of escaped semicolons in vmsify.
In theory, a Unix-format filespec can contain a semicolon and thus
need to be escaped when converted to a VMS-format filespec.  But
a much more common use case is a filespec that has a version
number despite being in Unix format.

So detect a semicolon that delimits a version specification and
pass it through but escape other semicolons.  This is apparently
what decc$to_vms does, so we're being consistent with the CRTL.