re_intuit_start(): eliminate checked_upto var
This var is supposed to record the latest place that stclass
had rejected, so if we try again, we don't start any earlier than that.
However, an assert showed that nothing in the test suite ever left
checked_upto > rx_origin on re-entry, so we could always use rx_origin
directly.

On failure, checked_upto is reset to HOPBACKc(endpos, start_shift)
where endpoint is the highest char that find_byclass() was asked to search
to. Now, I think that this is a logical error; it should have been
HOPBACKc(endpos, cl_l) or similar; i.e. just hop back the number of chars
equal to the length of the character class; using start_shift just makes
checked_upto artificially too small.

But in either case, consider the following more formal analysis (the
arithmetic below is in terms of chars).

Assume initially at least, that checked_upto <= rx_origin. The question is
whether we can ever end up with checked_upto "overtaking" rx_origin.

Where there is an anchored substring or ml_anch, we have:

    endpos = r_origin + cl_l;
    on failure,
        checked_upto = endpos - (start_shift or cl_l)

    since start_shift should be => cl_l, in either case we end up with
    checked_upto <= rx_origin.

Where there is only a floating substring, we have:

    endpos = r_origin - start_shift + cl_l;
    on failure,
        checked_upto = endpos - (start_shift or cl_l)

    again, since start_shift should be => cl_l, in either case we end up with
    checked_upto <= rx_origin.

Where there are no substrings, we return on failure; so the updating
checked_upto is irrelevant.

On success we return, so again the updating of checked_upto is irrelevant.