regen/regcharclass.pl: Generate correct macro instead of skipping
It makes no sense to check for length safeness for The macros generated
by this program which take a single UV code point as a parameter. Prior
to this patch, it would skip trying to generate them if asked.  But,
because of the way things are structured, that means that if you need
just this and the safe versions, you can't do it so easily.  What this
commit does is generate the cp macro if requested even if the 'safe'
version of other macros are also requested.