regcomp.c: Don't read past string-end
In doing an audit of regcomp.c, and experimenting using
Encode::_utf8_on(), I found this one instance of a regen/regcharclass.pl
macro that could read beyond the end of the string if given malformed
UTF-8.  Hence we convert to use the 'safe' form.  There are no other
uses of the non-safe version, so don't need to generate them.