tidy up kill0.t and kill0_child
The previous commit added some tests to kill0.t, and added the auxiliary
file kill0_child. Tidy up the new code to better match normal standards.
In particular, improve the format, grammar and clarity of the comments,
and replace q|...| with "..." where appropriate.
Also, make the temporary filename a variable, and prefix it with "tmp-",
so that if gets left around for any reason, it's more obvious that it's
just an extraneous temporary file.

(I haven't actually tested this commit on win32)