mktables: Fix overlooked in-line table defns code
Commit eb0925341cc65ce6ce57503ec0ab97cdad39dc98 introduced the idea of a
pseudo-directory as a way to store table definitions in-line in
Heavy.pl, but conform to the expectations of the code in regard to
objects being files within directories.  This kept the needed changes to
a minimum.  The code changed by the current commit  was overlooked then
as something that also needed to change, because there are no current
instances of it needing to.  But this could change with future Unicode
versions, or as in the next few commits, in extending the in-line
definitions.